Added corrections on story response code & types (see project website)

+++ payment/src/main/java/ca/ulaval/glo4002/payment/api/dto/RequestPaymentDto.java
@@ -1,16 +1,14 @@
 package ca.ulaval.glo4002.payment.api.dto;
 
-import java.math.BigDecimal;
-
 import ca.ulaval.glo4002.payment.domain.payment.PaymentMethod;
 
 public class RequestPaymentDto {
 
   private Long clientId;
-  private BigDecimal amount;
+  private float amount;
   private PaymentMethod paymentMethod;
 
-  public RequestPaymentDto(Long clientId, BigDecimal amount, PaymentMethod paymentMethod) {
+  public RequestPaymentDto(Long clientId, float amount, PaymentMethod paymentMethod) {
     this.clientId = clientId;
     this.amount = amount;
     this.paymentMethod = paymentMethod;
@@ -24,11 +22,11 @@ public class RequestPaymentDto {
     this.clientId = clientId;
   }
 
-  public BigDecimal getAmount() {
+  public float getAmount() {
     return amount;
   }
 
-  public void setAmount(BigDecimal amount) {
+  public void setAmount(float amount) {
     this.amount = amount;
   }
